requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.23.0
pillow>=9.0.0
opencv-python>=4.6.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
